requirements.txt
Django==1.9.6
django-picklefield==0.3.2
jsonfield